test: updating foundry test

+++ test/Treasury.t.sol
@@ -18,7 +18,7 @@ contract BaseSetup is Test {
     address internal dev;
     uint256 internal initialMint = 10_000_000_000e18;
     uint256 internal largeApproval = 100_000_000_000_000e18;
-    uint256 epochLen = 100;
+    uint32 epochLen = 100;
 
     function setUp() public virtual {
         utils = new Utils();
